Tests sorted. Tested and repaired additional methods in PIResponse.
